refactor: moved rekrutmen-card.tsx and use it on tim-saya page
