+++ src/hooks/useMyRekrutmen.ts
@@ -2,14 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { getToken } from "@/lib/cookies";
 import api from "@/lib/api";
 
-export interface Rekrutmen {
-  rekrutmen_id: string; // Updated to match the url usage /tim-saya/[rekrutmen_id]
-  role: string;
-  kegiatan: string;
-  tanggal_mulai: string;
-  tanggal_selesai: string;
-  fee: number;
-}
+import { Rekrutmen } from "@/hooks/useRekrutmen";
 
 export function useMyRekrutmen() {
   return useQuery({
